Fix writer data race and file-event counter overflow
WriterManager's per-trace-type StringBuilders and rotating filepath
fields were mutated without synchronization. Every trace type except
filesystem is written only from the single ETW processing thread, but:

  - the filesystem type is also written from the FilesystemHandlers flush
    timer (a thread-pool thread) when a deferred filename finally resolves
  - the snapshot types are written from their own snapper threads

StringBuilder is not thread-safe and FlushWrite rotates the per-type
filepath then compresses/deletes the old file, so two threads racing on
the filesystem path could corrupt CSV output, throw, double-compress, or
drop buffered events.

Guard each append + flush sequence with a per-trace-type lock. The
reconnect gate (ResumeGate.Wait) is taken before the lock so a long
reconnect can never pin it. FlushWrite is split into a public
gate+lock wrapper and a lock-held FlushWriteLocked core.

Also widen file_event_counter (and the persisted/reloaded LastFileEvent)
from int to long. The tracer runs for days at high event rates, so the
int counter could overflow ~2.1B events and the reload cast truncated
the persisted value.

+++ IOTracesCORE/cloudstorage/ObjectStorageHandler.cs
@@ -17,7 +17,7 @@ namespace IOTracesCORE.cloudstorage
         private R2Client r2Client;
         private ConcurrentQueue<string> uploadQueue = new();
         public static int UploadedFiles = 0;
-        public static int LastFileEvent;
+        public static long LastFileEvent;
 
         // ── Local upload buffering ────────────────────────────────────────────
         // Rather than uploading every flushed trace chunk individually (which
